+++ changelog.md
@@ -5,6 +5,7 @@
 - Fixed padding of logo in top banner on Settings screen
 - Fixed loading spinner for track 
 - Fixed hidden progress bar on Now Playing preview card
+- Added scroll-down to refresh to Home Screen
 
 
 # v2.0